move modules into correct folder
